requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic[email]==2.9.2   # or your FastAPI-compatible version

